micro.ui minor fixes plane deform demo complex number mul benchmark

+++ extras/test.linux.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 export "CMPL_HOME=$(dirname "$(dirname "$(readlink -f "$0")")")"
 echo "cmpl home is: $CMPL_HOME"
@@ -50,25 +50,27 @@ printf '%s\n' $DOC_FILES | while read file; do
 done
 
 # dump symbols, assembly, syntax tree and global variables
-TEST_FLAGS="$(echo -X+steps-stdin-offsets -asm/m/n/s -debug/g "$CMPL_HOME/cmplStd/test/test.ci")"
+TEST_FLAGS="$(echo -X+steps-stdin-offsets -asm/n/s -debug/g "$CMPL_HOME/cmplStd/test/test.ci")"
 $BIN_WCC/cmpl -log/d "$BIN_WCC.ci" $TEST_FLAGS
 if ! $BIN/cmpl -log/d "$BIN.ci" $TEST_FLAGS; then
 	echo "main test failed: $TEST_FLAGS"
 	exit 1
 fi
 
-#exit 0
-
 # dump symbols, documentation, assembly, syntax tree and global variables (to be compared with previous version to test if the code is generated properly)
-$BIN/cmpl -X+steps+fold+fast-stdin-glob-offsets -debug/G/M -api/A/m/d/p -asm/n/s -ast -doc -log/d/15 "extras/dump/test.dump.ci" -dump.ast.xml "extras/dump/test.dump.xml" "cmplStd/test/test.ci"
+$BIN/cmpl -X+steps+fold+fast-stdin-glob-offsets -debug/G/M -api/A/d/p -asm/n/s -ast -doc -log/d/15 "extras/dump/test.dump.ci" -dump.ast.xml "extras/dump/test.dump.xml" "cmplStd/test/test.ci"
 # dump symbols, documentation, assembly, syntax tree and global variables (to be compared with previous version to test if the code is generated properly)
-$BIN/cmpl -X+steps+fold+fast-stdin-glob-offsets -debug/G/M -api/A/m/d/p -asm/n/s -ast -doc -use -log/d "extras/dump/libs.dump.ci" "$BIN/libFile.so" "$BIN/libGfx.so"
+$BIN/cmpl -X+steps+fold+fast-stdin-glob-offsets -debug/G/M -api/A/d/p -asm/n/s -ast -doc -use -log/d "extras/dump/libs.dump.ci" "$BIN/libFile.so" "$BIN/libGfx.so"
 # dump profile data in text format including function tracing
-$BIN/cmpl -X-stdin+steps -profile/t/P/G/M -api/A/m/d/p -asm/g/n/s -ast/t -doc -use -log/d/15 "extras/dump/test.prof.ci" "cmplStd/test/test.ci"
+$BIN/cmpl -X-stdin+steps -profile/t/P/G/M -api/A/d/p -asm/g/n/s -ast/t -doc -use -log/d/15 "extras/dump/test.prof.ci" "cmplStd/test/test.ci"
 # dump profile data in json format
-$BIN/cmpl -X-stdin-steps -profile/t/P/G/M -api/A/m/d/p -asm/g/n/s -ast/t -doc -use -dump.json "extras/dump/test.prof.json" "cmplStd/test/test.ci"
+$BIN/cmpl -X-stdin-steps -profile/t/P/G/M -api/A/d/p -asm/g/n/s -ast/t -doc -use -dump.json "extras/dump/test.prof.json" "cmplStd/test/test.ci"
 
-#exit 0
+read -rsn1 -p "Compilation finished, press enter to run tests"
+if [ -n "$REPLY" ]; then
+	echo
+	exit 0
+fi
 
 DUMP_FILE=$BIN.antlr.dump.ci
 rm "$DUMP_FILE"
@@ -96,6 +98,7 @@ do
 		echo "**** cannot run test: $file"
 		continue
 	fi
+	printf "**** test: %s\\r" "$file"
 	if ! $BIN/cmpl -X-stdin+steps -run -log/a/d "$DUMP_FILE" "$BIN/libFile.so" "$BIN/libGfx.so" "$file"; then
 		echo "****** test failed: $file"
 	else
@@ -110,6 +113,7 @@ do
 		echo "**** cannot run test: $file"
 		continue
 	fi
+	printf "**** test: %s\\r" "$file"
 	if ! $BIN/cmpl -X-stdin+steps -run -log/a/d "$DUMP_FILE" "$BIN/libFile.so" "$BIN/libOpenGL.so" "$file"; then
 		echo "****** test failed: $file"
 	else
